+++ src/App/Bundle/MainBundle/Entity/Vote/Vote.php
@@ -1,0 +1,134 @@
+<?php
+
+namespace App\Bundle\MainBundle\Entity\Vote;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * Vote
+ *
+ * @ORM\Table(name="vote")
+ * @ORM\Entity(repositoryClass="App\Bundle\MainBundle\Entity\Vote\VoteRepository")
+ */
+class Vote
+{
+    /**
+     * The unique identifier
+     *
+     * @var int
+     *
+     * @ORM\Column(name="id", type="integer")
+     * @ORM\Id
+     * @ORM\GeneratedValue(strategy="AUTO")
+     */
+    private $identifier;
+
+    /**
+     * The unique identifier of the object voted
+     *
+     * @var int
+     *
+     * @ORM\Column(name="object", type="integer", nullable=false)
+     */
+    private $object;
+
+    /**
+     * The username of the user voting
+     *
+     * @var string
+     *
+     * @ORM\Column(name="username", type="string", nullable=false)
+     */
+    private $username;
+
+    /**
+     * The date at which the vote has been posted
+     *
+     * @var \DateTime
+     *
+     * @ORM\Column(name="created_at", type="datetime", nullable=false)
+     */
+    private $createdAt;
+
+    /**
+     * Get identifier
+     *
+     * @return int
+     */
+    public function getIdentifier()
+    {
+        return $this->identifier;
+    }
+
+    /**
+     * Get object
+     *
+     * @return int
+     */
+    public function getObject()
+    {
+        return $this->object;
+    }
+
+    /**
+     * Set object
+     *
+     * @param int $object
+     *
+     * @return self
+     */
+    public function setObject($object)
+    {
+        $this->object = $object;
+
+        return $this;
+    }
+
+    /**
+     * Get username
+     *
+     * @return string
+     */
+    public function getUsername()
+    {
+        return $this->username;
+    }
+
+    /**
+     * Set username
+     *
+     * @param string $username
+     *
+     * @return self
+     */
+    public function setUsername($username)
+    {
+        $this->username = $username;
+
+        return $this;
+    }
+
+    /**
+     * Get createdAt
+     *
+     * @return \DateTime
+     */
+    public function getCreatedAt()
+    {
+        return $this->createdAt;
+    }
+
+    /**
+     * Set createdAt
+     *
+     * @param \DateTime $createdAt
+     *
+     * @return self
+     */
+    public function setCreatedAt(\DateTime $createdAt)
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+}