+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 ### Added
 - Number of upvote on each world record
+- Ability to upvote a world record
 
 ### Changed